# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Open Arms nonprofit management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Dashboard Summary ──────────────────────────────────────────────
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incident reports
      parameters:
        - name: residentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: severity
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncidentReport"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create an incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentBody"
      responses:
        "201":
          description: Incident report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentReport"
  /incidents/{id}:
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentBody"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentReport"

  # ── Supporters ───────────────────────────────────────────────────────────
...
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: supporterId
          in: query
          schema:
            type: ["integer", "null"]
        - name: campaign
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated donations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDonations"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Record a new donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationBody"
      responses:
        "201":
          description: Donation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get donation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    patch:
      operationId: updateDonation
      tags: [donations]
      summary: Update a donation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonationBody"
      responses:
        "200":
          description: Updated donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerBody"
      responses:
        "201":
          description: Partner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get a partner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerBody"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"

  # ── Analytics ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, staff, donor]
        email:
          type: ["string", "null"]
      required: [id, username, displayName, role]

    # ── Contact Form ──────────────────────────────────────────────────────
...
    IncidentReport:
      type: object
      properties:
        id:
          type: integer
        residentId:
          type: ["integer", "null"]
        safehouseId:
          type: ["integer", "null"]
        incidentDate:
          type: string
        incidentType:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        reportedBy:
          type: ["string", "null"]
        resolved:
          type: boolean
        createdAt:
          type: string
      required: [id, incidentDate, severity, resolved, createdAt]
    CreateIncidentBody:
      type: object
      properties:
        residentId:
          type: ["integer", "null"]
        safehouseId:
          type: ["integer", "null"]
        incidentDate:
          type: string
        incidentType:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        reportedBy:
          type: ["string", "null"]
        resolved:
          type: boolean
          default: false
      required: [incidentDate, severity]
    UpdateIncidentBody:
      type: object
      properties:
        incidentType:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        reportedBy:
          type: ["string", "null"]
        resolved:
          type: ["boolean", "null"]

    # ── Supporter ─────────────────────────────────────────────────────────
...
    Donation:
      type: object
      properties:
        id:
          type: integer
        supporterId:
          type: ["integer", "null"]
        supporterName:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        donationType:
          type: string
          enum: [cash, bank_transfer, check, in_kind, online, recurring]
        campaign:
          type: ["string", "null"]
        channel:
          type: ["string", "null"]
        donatedAt:
          type: string
        receiptIssued:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, amount, currency, donationType, donatedAt, receiptIssued, createdAt]
    PaginatedDonations:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    CreateDonationBody:
      type: object
      properties:
        supporterId:
          type: ["integer", "null"]
        amount:
          type: number
        currency:
          type: string
          default: IDR
        donationType:
          type: string
          enum: [cash, bank_transfer, check, in_kind, online, recurring]
        campaign:
          type: ["string", "null"]
        channel:
          type: ["string", "null"]
        donatedAt:
          type: string
        receiptIssued:
          type: boolean
          default: false
        notes:
          type: ["string", "null"]
      required: [amount, currency, donationType, donatedAt]
    UpdateDonationBody:
      type: object
      properties:
        amount:
          type: ["number", "null"]
        campaign:
          type: ["string", "null"]
        channel:
          type: ["string", "null"]
        receiptIssued:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
...
    Partner:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        activeStatus:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, activeStatus, createdAt]
    CreatePartnerBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        activeStatus:
          type: boolean
          default: true
        notes:
          type: ["string", "null"]
      required: [name]
    UpdatePartnerBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        activeStatus:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]

    # ── Analytics ─────────────────────────────────────────────────────────
...