# lib/api-zod/src/generated/types/createDonationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Arms nonprofit management system API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDonationBodyDonationType } from "./createDonationBodyDonationType";

export interface CreateDonationBody {
  /** @nullable */
  supporterId?: number | null;
  amount: number;
  currency: string;
  donationType: CreateDonationBodyDonationType;
  /** @nullable */
  campaign?: string | null;
  /** @nullable */
  channel?: string | null;
  donatedAt: string;
  receiptIssued?: boolean;
  /** @nullable */
  notes?: string | null;
}